requirements.txt
fastapi
uvicorn[standard]
psycopg2
dotenv
sqlalchemy
typing
redis
requests
google-genai
pillow
python-multipart
alembic